Split cache out of planner
